added explosion and monster fadeout on kill

+++ project/Assets/TeamFiles/Scripts/GameScripts/MonsterScript.cs
@@ -1,6 +1,7 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.Runtime.InteropServices.WindowsRuntime;
+using Unity.VisualScripting;
 using UnityEngine;
 using UnityEngine.AI;
 using UnityEngine.UI;
@@ -30,6 +31,8 @@ public class MonsterScript : MonoBehaviour
     [SerializeField]
     private AudioSource deathAudioSource;
 
+    private bool startFadeOut;
+
 
     // Start is called before the first frame update
     void Start()
@@ -55,6 +58,7 @@ public class MonsterScript : MonoBehaviour
             spriteRenderer.color = new Color(spriteRenderer.color.r, spriteRenderer.color.g, spriteRenderer.color.b, 1f - fadeTimer);
             fadeTimer -= Time.deltaTime;
         }
+        StartFadeOut();
         
         //if (moveTimer <= 0)
         //{
@@ -83,6 +87,24 @@ public class MonsterScript : MonoBehaviour
             deathAudioSource.Play();
             Debug.Log("This monster trigger was hit by: Projectile");
             Destroy(other.gameObject);
+
+            spriteRenderer.color += new Color(0f,0f,0f,1f);
+            fade = true;
+            fadeTimer = 1.0f;
+            startFadeOut = true;
+        }
+    }
+
+    void StartFadeOut()
+    {
+        if(!startFadeOut) return;
+        if (fadeTimer > 0)
+        {
+            spriteRenderer.color = new Color(spriteRenderer.color.r, spriteRenderer.color.g, spriteRenderer.color.b, 0f + fadeTimer);
+            fadeTimer -= Time.deltaTime;
+        }
+        else if (fade && fadeTimer <= 0)
+        {
             Destroy(gameObject);
             playerScript.Kill();
         }
